+++ src/main/java/com/cashujdk/nut23/PostMintQuoteBolt11Request.java
@@ -1,15 +1,13 @@
 package com.cashujdk.nut23;
 
-import java.math.BigInteger;
-
 import com.cashujdk.nut04.PostMintQuoteRequest;
 
 public class PostMintQuoteBolt11Request extends PostMintQuoteRequest {
 
-    public BigInteger amount;
+    public long amount;
     public String description;
 
-    public PostMintQuoteBolt11Request(String unit, BigInteger amount, String description) {
+    public PostMintQuoteBolt11Request(String unit, long amount, String description) {
         super(unit);
         this.amount = amount;
         this.description = description;
