+++ src/main/java/com/cashujdk/nut23/PostMintQuoteBolt11Response.java
@@ -1,16 +1,14 @@
 package com.cashujdk.nut23;
 
-import java.math.BigInteger;
-
 import com.cashujdk.nut04.PostMintQuoteResponse;
 
 public class PostMintQuoteBolt11Response extends PostMintQuoteResponse {
 
-    public BigInteger amount;
+    public long amount;
     public String state;
-    public BigInteger expiry;
+    public long expiry;
 
-    public PostMintQuoteBolt11Response(String unit, String quote, String request, BigInteger amount, String state, BigInteger expiry) {
+    public PostMintQuoteBolt11Response(String unit, String quote, String request, long amount, String state, long expiry) {
         super(quote, request, unit);
         this.amount = amount;
         this.state = state;
